+++ app/controllers/products_controller.rb
@@ -164,6 +164,7 @@ class ProductsController < ApplicationController
       @possessions = current_user.possessions.where(product_id: @product.id, product_variant_id: nil)
       @bookmark = current_user.bookmarks.find_by(product_id: @product.id, product_variant_id: nil)
       @prev_owned = current_user.prev_owneds.find_by(product_id: @product.id, product_variant_id: nil)
+      @note = current_user.notes.find_by(product_id: @product.id, product_variant_id: nil)
       @setups = current_user.setups.includes(:possessions)
     end