+++ app/models/note.rb
@@ -1,0 +1,5 @@
+class Note < ApplicationRecord
+  belongs_to :product, optional: true
+  belongs_to :product_variant, optional: true
+  belongs_to :user
+end