+++ db/migrate/20240318144254_add_user_id_to_notes.rb
@@ -1,0 +1,5 @@
+class AddUserIdToNotes < ActiveRecord::Migration[7.1]
+  def change
+    add_column :notes, :user_id, :bigint, null: false
+  end
+end