+++ db/migrate/20240318160634_make_notes_for_product_per_user_unique.rb
@@ -1,0 +1,5 @@
+class MakeNotesForProductPerUserUnique < ActiveRecord::Migration[7.1]
+  def change
+    add_index :notes, [:user_id, :product_id, :product_variant_id]
+  end
+end